feat: 条件式ビルダをd1に移動
